ViTsolution/vit_ai_detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
tensorboard>=2.12.0
kagglehub>=0.1.0